+++ package.json
@@ -15,7 +15,8 @@
     "prettier": "^3.0.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-i18next": "^13.2.2"
+    "react-i18next": "^13.2.2",
+    "react-router-dom": "6"
   },
   "devDependencies": {
     "@types/react": "^18.2.15",
